feat(mobile): implemented get all listings

+++ mobile/lib/src/features/dashboard/data/models/payment_information.dart
@@ -17,7 +17,7 @@ class PaymentInformationModel extends PaymentInformation {
       updatedAt: DateTime.parse(json['updatedAt']),
       currency: json['currency'],
       paymentFrequency: json['paymentFrequency'],
-      cost: json['cost'],
+      cost: json['cost'].toDouble(),
       negotiable: json['negotiable'],
     );
   }
